refactor(frontend): extract JobsTab — final tab, App.jsx split done (17/N)
Last tab extraction. JobsTab moved to tabs/JobsTab.jsx with ~40 props
covering filters, search, sort, pinned states/cities, and per-card slice
of state for the JobCard list.

═══════════════════════════════════════════════════════════════════
  APP.JSX SPLIT COMPLETE
═══════════════════════════════════════════════════════════════════

Final tally after 17 PRs (#69-79 + #80-84):
  Before: 3886 LOC in one file
  After:  1805 LOC in App.jsx (orchestrator state + nav + tab routing)
          + ~2700 LOC distributed across:
            lib/{api,theme,format,jobConstants}.js          ~230 LOC
            components/{BrandLogo,LogoImg,ScrapeProgressBar,
                       SetupPanel,Pill,Chips,MRow,VaultRow,JobCard}.jsx  ~880 LOC
            tabs/{Monitor,Pipeline,Trends,Analytics,Companies,
                  Rare,Tracker,Vault,Jobs}Tab.jsx          ~1900 LOC

App.jsx is now scannable in one screen of context window.
Every feature has a single, predictable home.

Vite build clean across all 17 incremental PRs. Bundle size unchanged
within rounding (~680KB minified / 190KB gzip — same as before).
